Criado logs e colocado algumas verificacoes

+++ import.py
@@ -16,6 +16,7 @@ from datetime import date
 ################################################################
 origem = "C:\\01Digital\\aes\\Import"
 destino_historico = "C:\\01Digital\\aes\\Import\\Historico"
+destino_logs = "C:\\01Digital\\aes\\Import\\logs"
 host_database = "localhost"
 user_database = "root"
 pass_database =""
@@ -23,11 +24,12 @@ banco_database ="importador"
 #prefixo usado nas querys de insert
 prefixo_sql = "insert into temp_table_nova(origem,ds_0800,destino,situacao,OK, "
 prefixo_sql += "NR, LO, CO, CO2, CO3, DSC, OU, data, data_banco,ano, mes,dia, hora,duracao,ddd, semana) values"
-#lista de 0800 que serao imortados
-permitidos = ['08007272196','08007272196','08007271196','08007290196','08007701196','08007272120','08007240120','08007231000','08007232425','08007241234'
-,'08007290196','08007262626','08007230120','08007242424','08007245678','08007273123','08007700480','08007700260','08007728626']
 #numero de inserts por vez
 buffer_insert = 100
+
+
+
+
 
 ################################################################
 
@@ -46,6 +48,14 @@ def conectar():
                       passwd=pass_database, # your password
                       db=banco_database) # name of the data base
 	return db
+def grava_log(mensagem):
+	if not os.path.exists(destino_logs):
+		os.makedirs(destino_logs)
+	log_name = time.strftime("%Y_%m_%d_") +"log.txt"
+	text_log_file = open(destino_logs+"\\" + log_name, "a")		
+	text_log_file.write("[" + str(time.strftime("%Y-%m-%d %H:%M:%S")) + "]" + mensagem)	
+	text_log_file.close()
+	
 
 def executa_query(query):
 	db = conectar()	
@@ -68,93 +78,103 @@ def get_query(query):
 	finally:
 		db.close()
 	return data	
-
+	
 #os tres metodos abaixo estão de acordo com minha regra de negocio
 #caso queira usasr o programa, acredito que seria aqui o lugar
 #para aplicar sua regra...
 #nao repara se parecer estranha... mas é um sistema legado que 
 #irá receber esses dados	
-def importar_compato(arquivo):	
+def importar_compacto(arquivo):	
 	with open(origem + "\\" + arquivo, 'r') as linhas:
 		#prepara as variaveis
 		sql_string=""
 		n=0;
+		total_registros = 0
 		virgula=""
-
-		for linha in linhas.readlines():
-			n +=1
+		erro_colunas = False
+		for linha in linhas.readlines():			
 			colunas = linha.split(";")
-			sql_string += virgula + "("
-			#origem da chamada
-			sql_string += "'" + colunas[0] +"',"
+			total_registros += 1
+			if len(colunas) == 15:				
+				n +=1
+				sql_string += virgula + "("
+				#origem da chamada
+				sql_string += "'" + colunas[0] +"',"
 
-			#0800 destino
-			sql_string += "'" + colunas[1] +"',"
+				#0800 destino
+				sql_string += "'" + colunas[1] +"',"
 
-			#destino do 0800
-			sql_string += "'" + colunas[2] +"',"
+				#destino do 0800
+				sql_string += "'" + colunas[2] +"',"
 
-			#situacao/status da chamada
-			sql_string += "'" + colunas[4] +"',"
-			sql_string += "'" + ('1' if colunas[4]=="OK" else '0') +"',"
-			sql_string += "'" + ('1' if colunas[4]=="NR" else '0') +"',"
-			sql_string += "'" + ('1' if colunas[4]=="LO" else '0') +"',"
-			sql_string += "'" + ('1' if colunas[4]=="CO" else '0') +"',"
-			sql_string += "'" + ('1' if colunas[4]=="CO2" else '0') +"',"
-			sql_string += "'" + ('1' if colunas[4]=="CO3" else '0') +"',"
-			sql_string += "'" + ('1' if colunas[4]=="DSC" else '0') +"',"
+				#situacao/status da chamada
+				sql_string += "'" + colunas[4] +"',"
+				sql_string += "'" + ('1' if colunas[4]=="OK" else '0') +"',"
+				sql_string += "'" + ('1' if colunas[4]=="NR" else '0') +"',"
+				sql_string += "'" + ('1' if colunas[4]=="LO" else '0') +"',"
+				sql_string += "'" + ('1' if colunas[4]=="CO" else '0') +"',"
+				sql_string += "'" + ('1' if colunas[4]=="CO2" else '0') +"',"
+				sql_string += "'" + ('1' if colunas[4]=="CO3" else '0') +"',"
+				sql_string += "'" + ('1' if colunas[4]=="DSC" else '0') +"',"
 
-			#o resto sera considerado como OU
-			if not colunas[4] in ["OK", "NR", "LO", "CO", "CO2", "CO3", "DSC"] :
-				sql_string += "'1',"
+				#o resto sera considerado como OU
+				if not colunas[4] in ["OK", "NR", "LO", "CO", "CO2", "CO3", "DSC"] :
+					sql_string += "'1',"
+				else:
+					sql_string += "'0',"			
+				
+				#data, apenas para exibir em um grafico... não sera computado como data
+				sql_string += "'" +colunas[5][0:2] + "/" + colunas[5][3:5] + "/" + "20" + colunas[5][6:8]  + "',"
+
+				#data_banco
+				sql_string += "'20" + colunas[5][6:8] + "-" +colunas[5][3:5] + "-" + colunas[5][0:2] + "',"
+
+
+				#campo ano, mes e dia
+				sql_string += "'20" + colunas[5][6:8] + "',"
+				sql_string += "'" +colunas[5][3:5]  + "',"
+				sql_string += "'" +colunas[5][0:2] + "',"
+
+				#hora com intervalo de 30 em 30
+				if colunas[5][13:15] > 30:
+					intervalo = '30'
+				else:
+					intervalo = '00'
+				sql_string += "'" + colunas[5][9:12]  + intervalo + "',"
+
+				#tempo da ligacao
+				sql_string += "'" + colunas[6] +"',"
+
+				#dd chamante
+				sql_string += "'" + colunas[8] +"',"
+
+				#dia da semana
+				sql_string += "'" + dia_semana[date(int("20" + colunas[5][6:8]) , int(colunas[5][3:5]) , int(colunas[5][0:2])).weekday()] +"'"
+				
+				sql_string += ")"
+
+				#aquela virguala basica so para ajeitar a query =D
+				virgula=","
+
+				#opa, da uma parada e faz o insert de acordo com 
+				#o tamanho do buffer
+				if n >= buffer_insert:		
+					#print prefixo_sql+" "+sql_string
+					executa_query(prefixo_sql+" "+sql_string)				
+					sql_string=""
+					n=0
+					virgula=""
 			else:
-				sql_string += "'0',"			
-			
-			#data, apenas para exibir em um grafico... não sera computado como data
-			sql_string += "'" +colunas[5][0:2] + "/" + colunas[5][3:5] + "/" + "20" + colunas[5][6:8]  + "',"
-
-			#data_banco
-			sql_string += "'20" + colunas[5][6:8] + "-" +colunas[5][3:5] + "-" + colunas[5][0:2] + "',"
+				erro_colunas=True
 
 
-			#campo ano, mes e dia
-			sql_string += "'20" + colunas[5][6:8] + "',"
-			sql_string += "'" +colunas[5][3:5]  + "',"
-			sql_string += "'" +colunas[5][0:2] + "',"
-
-			#hora com intervalo de 30 em 30
-			if colunas[5][13:15] > 30:
-				intervalo = '30'
-			else:
-				intervalo = '00'
-			sql_string += "'" + colunas[5][9:12]  + intervalo + "',"
-
-			#tempo da ligacao
-			sql_string += "'" + colunas[6] +"',"
-
-			#dd chamante
-			sql_string += "'" + colunas[8] +"',"
-
-			#dia da semana
-			sql_string += "'" + dia_semana[date(int("20" + colunas[5][6:8]) , int(colunas[5][3:5]) , int(colunas[5][0:2])).weekday()] +"'"
-			
-			sql_string += ")"
-
-			#aquela virguala basica so para ajeitar a query =D
-			virgula=","
-
-			#opa, da uma parada e faz o insert de acordo com 
-			#o tamanho do buffer
-			if n >= buffer_insert:		
-				#print prefixo_sql+" "+sql_string
-				executa_query(prefixo_sql+" "+sql_string)				
-				sql_string=""
-				n=0
-				virgula=""
+		if erro_colunas:
+			grava_log("Arquivo " + arquivo + " com colunas incorretas. Eram esperadas 15 colunas.\n")
 
 		#caso tenha sobrado alguma coisa para processar
 		if sql_string != "":		
 			executa_query(prefixo_sql+" "+sql_string)
+		return total_registros
 
 def importar_completo(arquivo):
 	with open(origem + "\\" + arquivo, 'r') as linhas:
@@ -162,13 +182,16 @@ def importar_completo(arquivo):
 		sql_string=""
 		n=0;
 		virgula=""
-
+		total_registros = 0
+		erro_colunas = False
+		#tabela que tem os 0800 cadastrados... nao vamos querer pegar tudo...
+		permitidos = get_query("select ds_0800 from banco_bp.tbl_grupo_0800 group by 1")
 		for linha in linhas.readlines():
-			if len(linha) > 10:				
-				n +=1
-
-				colunas = linha.split("|")				
-				if "0800" + colunas[1].split("0800")[1] in permitidos:
+			colunas = linha.split("|")	
+			total_registros += 1		
+			if len(colunas) == 10:				
+				n +=1							
+				if any("0800" + colunas[1].split("0800")[1] in t_0800 for t_0800 in permitidos):
 
 					sql_string += virgula + "("
 					#origem da chamada
@@ -238,10 +261,16 @@ def importar_completo(arquivo):
 						sql_string=""
 						n=0
 						virgula=""
+			else:
+				erro_colunas=True
+
+		if erro_colunas:
+			grava_log("Arquivo " + arquivo + " com colunas incorretas. Eram esperadas 10 colunas.\n")
 
 		#caso tenha sobrado alguma coisa para processar
 		if sql_string != "":						
 			executa_query(prefixo_sql+" "+sql_string)
+		return total_registros
 
 def post_processar():
 	#para evitar alguma catastrofe eu movo os dados para uma tabela
@@ -297,7 +326,7 @@ def compactar(arquivo):
 	#que o cliente envie duas vezes o mesmo arquivo
 	#na pratica um vai anular o outro, mas vao ficar os dois no
 	#historico...
-	prefixo = time.strftime("%Y%m%d-%H%M%S") + "_" + arquivo + '.zip'
+	prefixo = arquivo.lower().replace(".txt","_") + time.strftime("%Y_%m_%d_%H_%M_%S") + '.zip'
 	zout = zipfile.ZipFile(origem + "\\" + prefixo, "w", zipfile.ZIP_DEFLATED) # <--- this is the change you need to make
 	zout.write(origem + "\\" + arquivo,os.path.basename(origem + "\\" + arquivo))
 	zout.close()
@@ -326,7 +355,7 @@ def historico(arquivo):
 		parte_data = 4
 
 	#faz o caminho para guardar organizado em ano/mes/dia
-	salvar_em = destino_historico+"\\" + partes_nome[parte_data][0:4]+"\\"+partes_nome[parte_data][4:6]+"\\"+partes_nome[parte_data][6:8]
+	salvar_em = destino_historico+"\\" + partes_nome[parte_data][0:4]+""+partes_nome[parte_data][4:6]+""+partes_nome[parte_data][6:8]
 	
 	#cria o diretorio caso nao exista
 	if not os.path.exists(salvar_em):
@@ -342,8 +371,8 @@ def historico(arquivo):
 #primeiro verifica se já esta rodando...
 #pode ser que demore o processo
 if not os.path.exists(running):
-	try:
-		print "Inicio    " + str(time.strftime("%H %M %S"))
+	try:		
+		executa_query("truncate table temp_table_nova")				
 		file_running = open(running,'w')  
 		file_running.close()
 
@@ -352,45 +381,59 @@ if not os.path.exists(running):
 			with open(job_list, 'r') as linhas:
 				for linha in linhas.readlines():
 					linha=linha.replace("\n","")
-					#se houver zip, apenas vou descompactar
-					#como nao sei oq vai sair de dentro
-					#no proximo passo procura os txt
-					if linha.endswith('.zip'):
-						descompactar(linha)					
+					if len(linha) > 5:
+						grava_log("Inicio do processamento do arquivo " + linha +".\n")					
+						#se houver zip, apenas vou descompactar
+						#como nao sei oq vai sair de dentro
+						#no proximo passo procura os txt
+						retorno_acao=""
 
-					#se tiver txt ja importa....
-					if linha.endswith('.txt'):
-						inicio_name = (linha + "_").split(' ', 1 )[0].split("_",1)[0]						
-						if inicio_name.isdigit():
-							#0800809494_20151115_sac.txt
-							importar_compato(linha)
-						else:
-							#Consulta_AES_CDRONE_SP01_20150905230000.txt
-							importar_completo(linha)
-						historico(linha)
+						if linha.upper().endswith('.ZIP'):
+							retorno_acao="[Descompactado]"
+							descompactar(linha)					
+
+						#se tiver txt ja importa....
+						
+						if linha.upper().endswith('.TXT'):
+							inicio_name = (linha + "_").split(' ', 1 )[0].split("_",1)[0]									
+							if inicio_name.isdigit():
+								#0800809494_20151115_sac.txt
+								retorno_acao = "[Linhas " + str(importar_compacto(linha)) + "]"
+							else:
+								#Consulta_AES_CDRONE_SP01_20150905230000.txt
+								retorno_acao = "[Linhas " + str(importar_completo(linha)) + "]"
+							historico(linha)
+						#faz carga em producao
+						#limpa as tabelas usadas no processo
+						#roda arquivo por arquivo para ficar mais leve de trabalhar com 
+						#tabelas menores....
+						post_processar()
+						grava_log("Fim do processamento do arquivo " + linha +". "+retorno_acao+"\n\n")	
 			
-			#faz carga em producao
-			#limpa as tabelas usadas no processo
-			post_processar()
-
-			#limpa a lista de jobs, tudo feito!
+			
+			
+	finally:	
+		#limpa a lista de jobs, tudo feito!
+		if os.path.exists(job_list):
 			os.remove(job_list)
-	finally:		
+
 		#remove o running e fica pronto para executar de novo
-		os.remove(running)
+		if os.path.exists(running):
+			os.remove(running)
 
 	#por fim verifica se tem novos arquivos para o proximo step
 	if not os.path.exists(job_list):
 		#recebe a lista de txts
-		get_txt_list = [f for f in os.listdir(origem) if f.endswith('.txt')]
+		get_txt_list = [f for f in os.listdir(origem) if f.upper().endswith('.TXT')]
 		#recebe a lista de zips
-		get_zip_list = [f for f in os.listdir(origem) if f.endswith('.zip')]
+		get_zip_list = [f for f in os.listdir(origem) if f.upper().endswith('.ZIP')]
 		#pode ser que tenhamos zips e txt ao mesmo tempo.
 
 
 		if len(get_txt_list) + len(get_zip_list) > 0:
+			grava_log("Arquivos encontrados para processamento.\n")	
 			text_file = open(job_list, "w")		
 	  		text_file.write("\n".join(get_txt_list) + "\n")
 	  		text_file.write("\n".join(get_zip_list) + "\n")
 			text_file.close()
-	print "Fim       " + str(time.strftime("%H %M %S"))
+	print "Fim "
